// src/types/user.type.ts
export interface IUser {
  id?: number;
  name?: string;
  mobile?: string;
  email?: string;
  password?: string;
  user_type?: string;
  role?: string;
  district_name?: string;
  division_id?: string;
  district_id?: string;
  upzilla_id?: string;
  union_id?: string;
  village?: string;
  bvc_registration_no?: string;
}